Prepare features for an IUCNN model

A wrapper function to prepare all default features included in IUCNN: geographic, biomes, climate, human footprint. If desired, bias features need to be calculated separately with ft_bias. For more control over feature preparation, you can use the iucnn_geography_features, iucnn_biome_features, iucnn_climate_features, iucnn_footprint_features functions.

iucnn_prepare_features(
  x,
  species = "species",
  lon = "decimallongitude",
  lat = "decimallatitude",
  type = c("geographic", "biomes", "climate", "humanfootprint"),
  download_folder = "feature_extraction"
)

Arguments

x
a data.frame of species occurrence records including three columns with species name, longitudinal coordinates and latitudinal coordinates (both decimal).

species
a character string. The name of the column with the species names.

lon
a character string. The name of the column with the longitude.

lat
a character string. The name of the column with the latitude.

type
character. The type of features to calculate. Possible options are “geographic”, “biome”, “climate”, “human footprint”.

download_folder
character string. The folder were to save the data used for feature extraction. Relative to the working directory. Set to NULL for the working directory

Value

a data.frame of features

Details

Without internet access, only geographic features are calculated,

See also

Examples

dat <- data.frame(species = c("A","B"),
                  decimallongitude = runif (200,10,15),
                  decimallatitude = runif (200,-5,5))

iucnn_prepare_features(dat)

#> Calculating geographic features.

#> [1] "Processing A"
#> [1] "Processing B"

#> Calculating biome features.

#> Warning: input biome polygon contained invalid geometries. Fixed using 'st_make_valid'

#> Calculating climate features.

#> Calculating human footprint features.

#> Downloading Human Footprint data from https://wcshumanfootprint.org/

#> Extracting_footprint_index for occurrence records

#> Summarizing information per species

#> Warning: Ignoring 55 records without data in the input raster

#> Adding missing grouping variables: `year`

#> # A tibble: 2 x 46
#>   species tot_occ uni_occ mean_lat mean_lon lat_range lon_range lat_hemisphere
#>   <chr>     <dbl>   <dbl>    <dbl>    <dbl>     <dbl>     <dbl>          <dbl>
#> 1 A          2.00    2.00  0.00767   0.0684     0.992     0.745              0
#> 2 B          2.00    2.00  0.00214   0.0692     0.993     0.733              0
#> # ... with 38 more variables: eoo <dbl>, aoo <dbl>, biome_7 <dbl>,
#> #   biome_1 <dbl>, biome_14 <dbl>, biome_2 <dbl>, biome_9 <dbl>, biome_3 <dbl>,
#> #   biome_13 <dbl>, biome_10 <dbl>, biome_4 <dbl>, biome_8 <dbl>,
#> #   biome_6 <dbl>, biome_11 <dbl>, biome_5 <dbl>, biome_99 <dbl>,
#> #   biome_12 <dbl>, biome_98 <dbl>, bio1 <dbl>, bio4 <dbl>, bio11 <dbl>,
#> #   bio12 <dbl>, bio15 <dbl>, bio17 <dbl>, range_bio1 <dbl>, range_bio4 <dbl>,
#> #   range_bio11 <dbl>, range_bio12 <dbl>, range_bio15 <dbl>, ...