Format IUCN Red List categories for IUCNN

Converting IUCN category labels into numeric categories required by train_iucnn.

prepare_labels(
  x,
  species = "species",
  labels = "labels",
  accepted_labels = c("CR", "EN", "VU", "NT", "LC"),
  level = "detail",
  threatened = c("CR", "EN", "VU")
)

Arguments

x
a data.frame or a list. If a data.frame, two columns with the species names and IUCN categories respectively. The column names are defined by the species and labels arguments. If a list, expecting the format as returned by rl_search.

species
a character string. The name of the column with the species names.

labels
a character string. The name of the column with the labels (assessment categories).

accepted_labels
a character string. The labels to be converted in to numeric values. Entries with labels not mentioned (e.g. "DD") will be removed. The numeric labels returned by the function will correspond to the order in this argument. For instance with default settings, "CR ->0, LC -> 4.

level
a character string. The level of output level detail. IF "detail" full iucn categories, if "broad" then 0 = Threatened, and 1 = Not threatened.

threatened
a character string. Only if level=="broad", Which labels to consider threatened.

Value

a data.frame with species names and numeric labels

Note

See vignette("Approximate_IUCN_Red_List_assessments_with_IUCNN") for a tutorial on how to run IUCNN.

Examples

data("training_labels")
prepare_labels(training_labels)

#> # A tibble: 889 x 2
#>    species                 labels
#>    <chr>                    <int>
#>  1 Liparis ambohimangana        1
#>  2 Liparis bautingensis         1
#>  3 Liparis bulbophylloides      2
#>  4 Liparis glaucescens          2
#>  5 Liparis goodyeroides         0
#>  6 Liparis henrici              2
#>  7 Liparis longicaulis          2
#>  8 Liparis longipetala          4
#>  9 Liparis petelotii            4
#> 10 Liparis philippinensis       4
#> # ... with 879 more rows